refactor(backend): extract shared types from pkg/loki to generic packages
- Move generic QueryResponse/AggregatedQueryResponse types from
  pkg/model/loki.go to pkg/model/query_response.go (Loki-only types remain)
- Extract Merger interface and MatrixMerger to pkg/merger package
- Extract TopologyInput struct and GetLabelsAndFilter to pkg/utils/queryparams
- Update pkg/loki to use type aliases for backward compatibility
- Update pkg/prometheus/query.go to import queryparams instead of loki
- Move matrix_merger_test.go to pkg/merger (unexported field access)
