+++ pkg/loki/merger.go
@@ -1,10 +1,9 @@
 package loki
 
 import (
-	"github.com/netobserv/network-observability-console-plugin/pkg/model"
+	"github.com/netobserv/network-observability-console-plugin/pkg/merger"
 )
 
-type Merger interface {
-	Add(from model.QueryResponseData) (model.ResultValue, error)
-	Get() *model.AggregatedQueryResponse
-}
+// Merger is an alias for merger.Merger for backwards compatibility within this package.
+// New code should use merger.Merger directly.
+type Merger = merger.Merger
