+++ pkg/merger/matrix_merger_test.go
@@ -1,4 +1,4 @@
-package loki
+package merger
 
 import (
 	"testing"
@@ -11,9 +11,13 @@ import (
 	"github.com/netobserv/network-observability-console-plugin/pkg/model"
 )
 
+func qrData(result model.ResultValue) model.QueryResponseData {
+	return model.QueryResponseData{Result: result}
+}
+
 func TestMatrixMerge(t *testing.T) {
 	now := pmodel.Now()
-	merger := NewMatrixMerger(100)
+	m := NewMatrixMerger(100)
 	baseline := pmodel.SampleStream{
 		Metric: pmodel.Metric{
 			"foo": "bar",
@@ -23,11 +27,11 @@ func TestMatrixMerge(t *testing.T) {
 			Value:     pmodel.SampleValue(42),
 		}},
 	}
-	_, err := merger.Add(qrData(model.Matrix{baseline}))
+	_, err := m.Add(qrData(model.Matrix{baseline}))
 	require.NoError(t, err)
 
 	// Different metric, different value pair => no dedup
-	_, err = merger.Add(qrData(model.Matrix{{
+	_, err = m.Add(qrData(model.Matrix{{
 		Metric: pmodel.Metric{
 			"foo":  "bar",
 			"foo2": "bar2",
@@ -41,7 +45,7 @@ func TestMatrixMerge(t *testing.T) {
 		}},
 	}}))
 	require.NoError(t, err)
-	result := merger.Get().Result.(model.Matrix)
+	result := m.Get().Result.(model.Matrix)
 	assert.Len(t, result, 2)
 	assert.Len(t, result[0].Values, 1)
 	assert.Equal(t, result[0].Values[0].Value, pmodel.SampleValue(54))
@@ -49,7 +53,7 @@ func TestMatrixMerge(t *testing.T) {
 	assert.Equal(t, result[1].Values[0].Value, pmodel.SampleValue(42))
 
 	// Same metrics in different order => no dedup
-	_, err = merger.Add(qrData(model.Matrix{{
+	_, err = m.Add(qrData(model.Matrix{{
 		Metric: pmodel.Metric{
 			"foo2": "bar2",
 			"foo":  "bar",
@@ -69,7 +73,7 @@ func TestMatrixMerge(t *testing.T) {
 		Values: baseline.Values,
 	}}))
 	require.NoError(t, err)
-	result = merger.Get().Result.(model.Matrix)
+	result = m.Get().Result.(model.Matrix)
 	assert.Len(t, result, 2)
 	assert.Len(t, result[0].Values, 1)
 	assert.Equal(t, result[0].Values[0].Value, pmodel.SampleValue(54))
@@ -77,7 +81,7 @@ func TestMatrixMerge(t *testing.T) {
 	assert.Equal(t, result[1].Values[0].Value, pmodel.SampleValue(168))
 
 	// Different timestamp => no dedup
-	_, err = merger.Add(qrData(model.Matrix{{
+	_, err = m.Add(qrData(model.Matrix{{
 		Metric: baseline.Metric,
 		Values: []pmodel.SamplePair{{
 			Timestamp: now.Add(time.Hour),
@@ -85,7 +89,7 @@ func TestMatrixMerge(t *testing.T) {
 		}},
 	}}))
 	require.NoError(t, err)
-	result = merger.Get().Result.(model.Matrix)
+	result = m.Get().Result.(model.Matrix)
 	assert.Len(t, result, 2)
 	assert.Len(t, result[0].Values, 2)
 	assert.Equal(t, result[0].Values[0].Value, pmodel.SampleValue(54))
@@ -94,7 +98,7 @@ func TestMatrixMerge(t *testing.T) {
 	assert.Equal(t, result[1].Values[0].Value, pmodel.SampleValue(168))
 
 	// no dedup
-	_, err = merger.Add(qrData(model.Matrix{{
+	_, err = m.Add(qrData(model.Matrix{{
 		// changed value => no dedup
 		Metric: baseline.Metric,
 		Values: []pmodel.SamplePair{{
@@ -122,7 +126,7 @@ func TestMatrixMerge(t *testing.T) {
 
 	// Different timestamp
 	require.NoError(t, err)
-	result = merger.Get().Result.(model.Matrix)
+	result = m.Get().Result.(model.Matrix)
 	assert.Len(t, result, 2)
 	assert.Len(t, result[0].Values, 2)
 	assert.Equal(t, result[0].Values[0].Value, pmodel.SampleValue(104))
@@ -133,7 +137,7 @@ func TestMatrixMerge(t *testing.T) {
 
 func TestMatrixLimitReached(t *testing.T) {
 	now := pmodel.Now()
-	merger := NewMatrixMerger(2)
+	m := NewMatrixMerger(2)
 
 	// Single entry => should not reach limit
 	first := pmodel.SampleStream{
@@ -145,12 +149,12 @@ func TestMatrixLimitReached(t *testing.T) {
 			Value:     pmodel.SampleValue(42),
 		}},
 	}
-	_, err := merger.Add(qrData(model.Matrix{first}))
+	_, err := m.Add(qrData(model.Matrix{first}))
 	require.NoError(t, err)
-	assert.False(t, merger.limitReached)
+	assert.False(t, m.limitReached)
 
 	// Another single entry => limit still not reached (even if total is 2)
-	_, err = merger.Add(qrData(model.Matrix{pmodel.SampleStream{
+	_, err = m.Add(qrData(model.Matrix{pmodel.SampleStream{
 		Metric: pmodel.Metric{
 			"foo": "bar",
 		},
@@ -160,10 +164,10 @@ func TestMatrixLimitReached(t *testing.T) {
 		}},
 	}}))
 	require.NoError(t, err)
-	assert.False(t, merger.limitReached)
+	assert.False(t, m.limitReached)
 
 	// 2 entries => limit reached
-	_, err = merger.Add(qrData(model.Matrix{pmodel.SampleStream{
+	_, err = m.Add(qrData(model.Matrix{pmodel.SampleStream{
 		Metric: pmodel.Metric{
 			"foo": "bar",
 		},
@@ -181,10 +185,10 @@ func TestMatrixLimitReached(t *testing.T) {
 		}},
 	}}))
 	require.NoError(t, err)
-	assert.True(t, merger.limitReached)
+	assert.True(t, m.limitReached)
 
 	// Another single entry => limit still reached
-	_, err = merger.Add(qrData(model.Matrix{pmodel.SampleStream{
+	_, err = m.Add(qrData(model.Matrix{pmodel.SampleStream{
 		Metric: pmodel.Metric{
 			"foo": "bar",
 		},
@@ -194,5 +198,5 @@ func TestMatrixLimitReached(t *testing.T) {
 		}},
 	}}))
 	require.NoError(t, err)
-	assert.True(t, merger.limitReached)
+	assert.True(t, m.limitReached)
 }
