+++ pkg/prometheus/query.go
@@ -3,15 +3,15 @@ package prometheus
 import (
 	"strings"
 
-	"github.com/netobserv/network-observability-console-plugin/pkg/loki"
 	"github.com/netobserv/network-observability-console-plugin/pkg/model/filters"
 	"github.com/netobserv/network-observability-console-plugin/pkg/utils/constants"
+	"github.com/netobserv/network-observability-console-plugin/pkg/utils/queryparams"
 	v1 "github.com/prometheus/client_golang/api/prometheus/v1"
 )
 
 type QueryBuilder struct {
 	aggregateKeyLabels map[string][]string
-	in                 *loki.TopologyInput
+	in                 *queryparams.TopologyInput
 	filters            filters.SingleQuery
 	orMetrics          []string
 	qRange             v1.Range
@@ -22,7 +22,8 @@ type Query struct {
 	PromQL string
 }
 
-func NewQuery(kl map[string][]string, in *loki.TopologyInput, qr *v1.Range, filters filters.SingleQuery, orMetrics []string) *QueryBuilder {
+// NewQuery creates a new PromQL QueryBuilder for the given topology input.
+func NewQuery(kl map[string][]string, in *queryparams.TopologyInput, qr *v1.Range, filters filters.SingleQuery, orMetrics []string) *QueryBuilder {
 	return &QueryBuilder{
 		aggregateKeyLabels: kl,
 		in:                 in,
@@ -158,14 +159,17 @@ func appendFilteredMetric(sb *strings.Builder, metric string, filters filters.Si
 	sb.WriteRune('}')
 }
 
+// GetLabelsAndFilter returns the label fields for grouping and an optional extra filter string.
+// For Prometheus, the "app" aggregate is a noop (it's only relevant for Loki stream selectors).
 func GetLabelsAndFilter(kl map[string][]string, aggregate, groups string) ([]string, string) {
 	if aggregate == "app" {
 		// ignore app: it's a noop aggregation needed for Loki, not relevant in promQL
 		return nil, ""
 	}
-	return loki.GetLabelsAndFilter(kl, aggregate, groups)
+	return queryparams.GetLabelsAndFilter(kl, aggregate, groups)
 }
 
+// QueryFilters builds a PromQL metric selector string with the given filters applied.
 func QueryFilters(metric string, filters filters.SingleQuery) string {
 	sb := strings.Builder{}
 	appendFilteredMetric(&sb, metric, filters)
