+++ pkg/prometheus/query_test.go
@@ -4,10 +4,10 @@ import (
 	"testing"
 	"time"
 
-	"github.com/netobserv/network-observability-console-plugin/pkg/loki"
 	"github.com/netobserv/network-observability-console-plugin/pkg/model/fields"
 	"github.com/netobserv/network-observability-console-plugin/pkg/model/filters"
 	"github.com/netobserv/network-observability-console-plugin/pkg/utils/constants"
+	"github.com/netobserv/network-observability-console-plugin/pkg/utils/queryparams"
 	v1 "github.com/prometheus/client_golang/api/prometheus/v1"
 	"github.com/stretchr/testify/assert"
 )
@@ -28,7 +28,7 @@ var kl = map[string][]string{
 }
 
 func TestBuildQuery_PromQLSimpleRateIgnoreApp(t *testing.T) {
-	in := loki.TopologyInput{
+	in := queryparams.TopologyInput{
 		Top:            "50",
 		RateInterval:   "2m",
 		DataField:      "Bytes",
@@ -48,7 +48,7 @@ func TestBuildQuery_PromQLSimpleRateIgnoreApp(t *testing.T) {
 }
 
 func TestBuildQuery_PromQLSimpleRateNoFilter(t *testing.T) {
-	in := loki.TopologyInput{
+	in := queryparams.TopologyInput{
 		Top:            "50",
 		RateInterval:   "2m",
 		DataField:      "Bytes",
@@ -68,7 +68,7 @@ func TestBuildQuery_PromQLSimpleRateNoFilter(t *testing.T) {
 }
 
 func TestBuildQuery_PromQLSimpleRateAndFilter(t *testing.T) {
-	in := loki.TopologyInput{
+	in := queryparams.TopologyInput{
 		Top:            "50",
 		RateInterval:   "2m",
 		DataField:      "Bytes",
@@ -93,7 +93,7 @@ func TestBuildQuery_PromQLSimpleRateAndFilter(t *testing.T) {
 }
 
 func TestBuildQuery_PromQLRateMultiFilter(t *testing.T) {
-	in := loki.TopologyInput{
+	in := queryparams.TopologyInput{
 		Top:            "50",
 		RateInterval:   "2m",
 		DataField:      "Bytes",
@@ -119,7 +119,7 @@ func TestBuildQuery_PromQLRateMultiFilter(t *testing.T) {
 }
 
 func TestBuildQuery_PromQLHistogramAverage(t *testing.T) {
-	in := loki.TopologyInput{
+	in := queryparams.TopologyInput{
 		Top:            "50",
 		RateInterval:   "2m",
 		DataField:      "TimeFlowRttNs",
@@ -144,7 +144,7 @@ func TestBuildQuery_PromQLHistogramAverage(t *testing.T) {
 }
 
 func TestBuildQuery_PromQLHistogramP99(t *testing.T) {
-	in := loki.TopologyInput{
+	in := queryparams.TopologyInput{
 		Top:            "50",
 		RateInterval:   "2m",
 		DataField:      "TimeFlowRttNs",
@@ -169,7 +169,7 @@ func TestBuildQuery_PromQLHistogramP99(t *testing.T) {
 }
 
 func TestBuildQuery_PromQLByDNSResponseCode(t *testing.T) {
-	in := loki.TopologyInput{
+	in := queryparams.TopologyInput{
 		Top:            "5",
 		RateInterval:   "1m",
 		DataField:      "DnsFlows",
@@ -189,7 +189,7 @@ func TestBuildQuery_PromQLByDNSResponseCode(t *testing.T) {
 }
 
 func TestBuildQuery_PromQLORMetrics(t *testing.T) {
-	in := loki.TopologyInput{
+	in := queryparams.TopologyInput{
 		Top:            "50",
 		RateInterval:   "2m",
 		DataField:      "Bytes",
